added event description

+++ Website.v2/src/views/web-events/web-events.ts
@@ -28,36 +28,37 @@ export default class WebEvents extends PageMixin(LitElement) {
         <div class="mb-3 events-list">
           <h1>Alle Buchungen</h1>
 
-
-          <table class="table">
-            <thead>
-              <tr>
-                <th scope="col">Titel</th>
-                <th scope="col">Start</th>
-                <th scope="col">Ende</th>
-                <th scope="col">Raum</th>
-                <th scope="col">Erstellt von</th>
-                <th scope="col"></th>
-              </tr>
-            </thead>
-            <tbody>
-              ${ this.events.map(event => { return html`
-              <tr>
-                <th scope="row">${event.title}</th>
-                <td>${this.formateDate(event.start.toDate())}</td>
-                <td>${this.formateDate(event.end.toDate())}</td>
-                <td>${event.room}</td>
-                <td>${event.createdFrom}</td>
-                ${this.user?.role === 'admin' || this.user?.id === event.createdFromId? html`
-                <td>
-                  <edit-event .event=${event} class="align-self-center me-3"></edit-event>
-                  <button type="button" class="btn btn-danger" @click=${() => this.deleteEvent(event.id)}>Löschen</button>
-                </td>
-                `: undefined}
-              </tr>
-            </tbody>
-            `})}
-          </table>
+          <div class="table-responsive">
+            <table class="table">
+              <thead>
+                <tr>
+                  <th scope="col">Titel</th>
+                  <th scope="col">Start</th>
+                  <th scope="col">Ende</th>
+                  <th scope="col">Raum</th>
+                  <th scope="col">Erstellt von</th>
+                  <th scope="col"></th>
+                </tr>
+              </thead>
+              <tbody>
+                ${ this.events.map(event => { return html`
+                <tr>
+                  <th scope="row">${event.title}</th>
+                  <td>${this.formateDate(event.start.toDate())}</td>
+                  <td>${this.formateDate(event.end.toDate())}</td>
+                  <td>${event.room}</td>
+                  <td>${event.createdFrom}</td>
+                  ${this.user?.role === 'admin' || this.user?.id === event.createdFromId? html`
+                  <td class="event-actions">
+                    <edit-event .event=${event} class="align-self-center me-3"></edit-event>
+                    <button type="button" class="btn btn-danger" @click=${() => this.deleteEvent(event.id)}>Löschen</button>
+                  </td>
+                  `: undefined}
+                </tr>
+                `})}
+              </tbody>
+            </table>
+          </div>
         </div>
       </div>
         `
@@ -71,7 +72,15 @@ export default class WebEvents extends PageMixin(LitElement) {
   stateChanged(state: IState): void {
     this.user = state.user;
     if (state.events.length > 0) {
-      this.events = state.events;
+      this.events = state.events.sort((a, b) => {
+        if (a.start > b.start) {
+          return 1;
+        } else if (a.start < b.start) {
+          return -1;
+        } else {
+          return 0;
+        }
+      });
     }
   }
   
@@ -84,7 +93,10 @@ export default class WebEvents extends PageMixin(LitElement) {
   }
 
   deleteEvent(id: string): void {
-    EventService.deleteEvent(id);
+    const r = confirm('Soll die Buchung wirklich gelöscht werden?');
+    if (r == true) {
+      EventService.deleteEvent(id);
+    }
   }
 
   formateDate(date: Date): string {
